another small shuffle
