+++ .gitignore
@@ -1,2 +1,2 @@
 ## Built binaries
-/mulery-server
+/mulery
